refactor/modularize controllers and routes
